user-service/requirements.txt
fastapi[all]
sqlalchemy[asyncio]
psycopg2-binary
asyncpg
alembic
passlib[argon2]
pyjwt
pytest
pytest-asyncio
anyio
httpx
